Add support for addition SDK mode flags.
The SDK mode now requires an optional flag in order
to override default localization responses.  When this
flag is set, the vehicle will be able to drive without
receiving correction localization input. This was previously
a default value.

In order to set speed and change lanes, the
ANKI_VEHICLE_SDK_OPTION_OVERRIDE_LOCALIZATION
must be passed when turning on SDK mode.